refactor: merge DataForm and Unterformular
