+++ src/model/mod.rs
@@ -29,12 +29,11 @@ use crate::model::requirements::Requires;
 use crate::profile::{FormField, FormReference, Profile, WithScriptsCode};
 
 pub mod data_catalogue;
-pub mod data_form;
+pub mod form;
 pub mod onkostar_editor;
 pub mod other;
 pub mod property_catalogue;
 pub mod requirements;
-pub mod unterformular;
 
 #[derive(Serialize, Deserialize, Debug, PartialEq)]
 #[serde(deny_unknown_fields)]
